Refactor: Move AMP URL detection related codes to a separate file
