+++ app/src/main/java/com/duckduckgo/app/browser/BrowserWebViewClient.kt
@@ -131,6 +131,7 @@ class BrowserWebViewClient @Inject constructor(
     private val subscriptions: Subscriptions,
     private val dnsResolver: CustomDnsResolver,
     private val sgWhitelistDao: SafeGazeWhitelistDao,
+    private val ampDetector: AmpDetector,
     spProvider: SharedPreferencesProvider
 ) : WebViewClient() {
 
@@ -519,13 +520,13 @@ class BrowserWebViewClient @Inject constructor(
         if (request.url.host == KAHF_GUARD_BLOCKED_URL || url.isDataUri()) return null
         val privateDnsMode = PrivateDnsLevel.getCurrentLevel(sharedPreferences)
         val privateDnsEnabled = privateDnsMode != PrivateDnsLevel.Off
-        val isAmpUrl = isAmpUrl(url)
+        val isAmpUrl = ampDetector.isAmpUrl(url)
 
         return runBlocking {
             withContext(dispatcherProvider.io()) {
                 try {
                     if (isAmpUrl) {
-                        extractOriginalUrlFromAmp(url).also {
+                        ampDetector.extractOriginalUrlFromAmp(url).also {
                             Timber.d("amLog AMP URL: $it")
                             url = it
                         }
@@ -747,43 +748,6 @@ class BrowserWebViewClient @Inject constructor(
             SAFE_BROWSING_THREAT_UNKNOWN -> "SAFE_BROWSING_THREAT_UNKNOWN"
             SAFE_BROWSING_THREAT_UNWANTED_SOFTWARE -> "SAFE_BROWSING_THREAT_UNWANTED_SOFTWARE"
             else -> "ERROR_OTHER"
-        }
-    }
-
-    private fun isAmpUrl(url: String) =
-        url.contains("cdn.ampproject.org") && url.toUri().host != "cdn.ampproject.org"
-
-    private fun extractOriginalUrlFromAmp(ampUrl: String): String {
-        return try {
-            val uri = ampUrl.toUri()
-
-            // Handle cdn.ampproject.org format
-            if (uri.host?.contains("cdn.ampproject.org") == true) {
-                val pathSegments = uri.pathSegments
-                val index = pathSegments.indexOf("v")
-                if (index != -1 && pathSegments.size > index + 2) {
-                    val isSecure = pathSegments[index + 1] == "s"
-                    val originalUrlPath = pathSegments.subList(index + 2, pathSegments.size).joinToString("/")
-                    val scheme = if (isSecure) "https://" else "http://"
-                    return scheme + originalUrlPath
-                }
-            }
-
-            // Handle Google AMP format
-            if (uri.host?.contains("google.com") == true && uri.path?.contains("/amp/") == true) {
-                val path = uri.path ?: return ampUrl
-                val prefix = "/amp/s/"
-                return if (path.contains(prefix)) {
-                    "https://" + path.substringAfter(prefix)
-                } else {
-                    "http://" + path.substringAfter("/amp/")
-                }
-            }
-
-            // Fallback
-            ampUrl
-        } catch (e: Exception) {
-            ampUrl // fallback
         }
     }
 }
